instanciate a mongokit connection explicetly in connection module
to allow to initiate a connection with the uri set in config and
avoiding cycling dependancies, we need to instanciate a Mogokit connection
explicitley in the connection module and use the connection object at
the top level of the module connection in the other modules which need
it.
modify the import according this new connection object location.