rework Type and SchemaDocument-based classes, work in progress
MK's SchemaDocument is clearly not meant to be a subdocument container
and keeps on generating issues when used this way.
e.g., A SchemaDocument cannot contain a CustomType, which is a pity.
All Mobyle2 classes using SchemaDocument as a superclass might need to
be re-worked on.
* mk_struct.py contains a workaround superclass and adapter which I need
to test further,
* test_type will now test that the CustomType can also be included in a
subdocument,
* type needs to be adapted, maybe to use mk_struct module
* service and m1_service_import are "type-related" work in progress.